// src/api/arrow_type.ts
export enum ArrowType {
    EMPTY = 0,
    RED_ARROW = 1,
    RED_SOURCE = 2,
    BLOCKER = 3,
    DELAY = 4,
    DETECTOR = 5,
    SPLITTER_1 = 6,
    SPLITTER_2 = 7,
    SPLITTER_3 = 8,
    RED_IMPULSE = 9,
    BLUE_ARROW = 10,
    BLUE_DIAGONAL_ARROW = 11,
    BLUE_SPLITTER_1 = 12,
    BLUE_SPLITTER_2 = 13,
    BLUE_SPLITTER_3 = 14,
    LOGIC_NOT = 15,
    LOGIC_AND = 16,
    LOGIC_XOR = 17,
    LOGIC_FLIP = 18,
    LOGIC_FLOP = 19,
    RANDOM = 20,
    BUTTON = 21,
    LEVEL_ARROW_22 = 22,
    LEVEL_ARROW_23 = 23,
    BRUH_BUTTON = 24,
}